rename AnnotationReplacerRector to RenameAnnotationRector
