+++ ecs.yml
@@ -81,7 +81,7 @@ parameters:
             # required for exact string match with "\"
             - 'packages/NodeTypeResolver/src/Php/AbstractTypeInfo.php'
             # example in description
-            - 'src/Rector/Annotation/AnnotationReplacerRector.php'
+            - 'src/Rector/Annotation/RenameAnnotationRector.php'
             - 'packages/NetteToSymfony/src/Event/EventInfosFactory.php'
 
         Symplify\CodingStandard\Sniffs\CleanCode\ForbiddenStaticFunctionSniff:
